feat: Convert connection string into list
I am having a lot of trouble using this Helm chart via
https://github.com/aws-ia/terraform-aws-eks-blueprints because the
rendering of the Helm chart fails due to the fact that there are strings
containing commas. I was hoping to be able to work around the problem by
escaping the commas, but that didn't work out in this case.